Remove clock from prompt. Cleanups.

+++ docker/install.zsh
@@ -5,5 +5,7 @@
 # Docker-desktop keeps changing the user .aws folder to link to
 # `/mnt/c/Users/kania/.aws`. This makes it point to the aws folder in my
 # private configurations repo.
-rm $HOME/.aws
-ln -s $HOME/projects/private/aws $HOME/.aws
+if [[ -d $HOME/projects/private/aws ]]; then
+  rm $HOME/.aws
+  ln -s $HOME/projects/private/aws $HOME/.aws
+fi
